refactor(workflow): split release.go into 5 focused files
